Рефакторинг: Разнесла на различные файлы
